+++ src/main/java/plugily/projects/buildbattle/arena/ArenaRegistry.java
@@ -19,6 +19,7 @@
 package plugily.projects.buildbattle.arena;
 
 import org.bukkit.Location;
+import org.bukkit.block.Biome;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.entity.Player;
@@ -34,6 +35,7 @@ import plugily.projects.buildbattle.arena.impl.SoloArena;
 import plugily.projects.buildbattle.arena.impl.TeamArena;
 import plugily.projects.buildbattle.arena.managers.plots.Plot;
 import plugily.projects.buildbattle.utils.Debugger;
+import plugily.projects.buildbattle.utils.ServerVersion.Version;
 
 import javax.annotation.Nullable;
 import java.util.*;
@@ -169,7 +171,10 @@ public class ArenaRegistry {
           for (String plotName : config.getConfigurationSection(s + "plots").getKeys(false)) {
             if (config.isSet(s + "plots." + plotName + ".maxpoint") && config.isSet(s + "plots." + plotName + ".minpoint")) {
               Location minPoint = LocationSerializer.getLocation(config.getString(s + "plots." + plotName + ".minpoint"));
-              Plot buildPlot = new Plot(arena, minPoint.getWorld().getBiome(minPoint.getBlockX(), minPoint.getBlockZ()));
+              Biome biome = Version.isCurrentHigher(Version.v1_15_R1) ?
+                      minPoint.getWorld().getBiome(minPoint.getBlockX(), minPoint.getBlockY(), minPoint.getBlockZ())
+                      : minPoint.getWorld().getBiome(minPoint.getBlockX(), minPoint.getBlockZ());
+              Plot buildPlot = new Plot(arena, biome);
               buildPlot.setCuboid(new Cuboid(minPoint, LocationSerializer.getLocation(config.getString(s + "plots." + plotName + ".maxpoint"))));
               buildPlot.fullyResetPlot();
               arena.getPlotManager().addBuildPlot(buildPlot);