+++ src/main/java/plugily/projects/buildbattle/handlers/ChatManager.java
@@ -21,10 +21,8 @@ package plugily.projects.buildbattle.handlers;
 import org.apache.commons.lang.StringUtils;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
-import org.bukkit.plugin.java.JavaPlugin;
 
 import pl.plajerlair.commonsbox.string.StringFormatUtils;
-import plugily.projects.buildbattle.Main;
 import plugily.projects.buildbattle.arena.impl.BaseArena;
 import plugily.projects.buildbattle.handlers.language.LanguageManager;
 import plugily.projects.buildbattle.utils.ServerVersion;