+++ src/main/java/plugily/projects/buildbattle/handlers/reward/RewardsFactory.java
@@ -53,6 +53,10 @@ public class RewardsFactory {
       return;
     }
     BaseArena arena = ArenaRegistry.getArena(player);
+    if (arena == null) {
+      return;
+    }
+
     ScriptEngine engine = new ScriptEngine();
     engine.setValue("player", player);
     engine.setValue("server", Bukkit.getServer());