+++ src/main/java/plugily/projects/buildbattle/arena/managers/plots/Plot.java
@@ -191,6 +191,9 @@ public class Plot {
             continue;
           }
           if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_16_R1)) {
+            Utils.sendPacket(p, Utils.getNMSClass("PacketPlayOutMapChunk").getConstructor(Utils.getNMSClass("Chunk"), int.class)
+                    .newInstance(chunk.getClass().getMethod("getHandle").invoke(chunk), 65535));
+          } else if (ServerVersion.Version.isCurrentEqualOrHigher(ServerVersion.Version.v1_15_R1)) {
             Utils.sendPacket(p, Utils.getNMSClass("PacketPlayOutMapChunk").getConstructor(Utils.getNMSClass("Chunk"), int.class, boolean.class)
                     .newInstance(chunk.getClass().getMethod("getHandle").invoke(chunk), 65535, false));
           } else {